[UPDATE] Home screen color design

+++ App/app/src/main/java/com/example/eventplanner/model/Product.java
@@ -1,7 +1,7 @@
 package com.example.eventplanner.model;
 
 public class Product extends ServiceProduct {
-    public Product(String name, double price, double discount, Boolean isAvailable, double grade, String category, int image) {
+    public Product(String name, int price, int discount, Boolean isAvailable, double grade, String category, int image) {
         super(name, price, discount, isAvailable, grade, category, image);
     }
 }
